feat: 🎸 Creates Account Management namespace
Creates and moves account management namespace and moves methods into it

- sdk.currentIdentity.freezeSecondaryKeys -> sdk.accountManagement.freezeSecondaryKeys
- sdk.currentIdentity.unfreezeSecondaryKeys  -> sdk.accountManagement.unfreezeSecondaryKeys
- sdk.currentIdentity.revokePermissions  -> sdk.accountManagement.revokePermissions
- sdk.currentIdentity.modifyPermissions  -> sdk.accountManagement.modifyPermissions
- sdk.currentIdentity.removeSecondaryKeys  -> sdk.accountManagement.removeSecondaryKeys
- sdk.currentIdentity.inviteAccount  -> sdk.accountManagement.inviteAccount
- sdk.getAccount  -> sdk.accountManagement.getAccount
- sdk.getAccounts  -> sdk.accountManagement.getAccounts
- sdk.getAccountBalance  -> sdk.accountManagement.getAccountBalance

BREAKING CHANGE: 🧨 Move account management methods
